requirements.txt
asgiref==3.6.0
Django==4.1.5
django-appconf==1.0.5
django-imagekit==4.1.0
mysqlclient==2.2.0
pilkit==2.0
Pillow==9.4.0
python-decouple==3.8
six==1.16.0
sqlparse==0.4.3
